requirements.txt
# Telegram бот
python-telegram-bot==20.7
python-dotenv==1.0.0

# Обработка изображений
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3

# IOPaint - готовая обертка над LaMa (упрощенная интеграция)
iopaint==1.3.3

# Утилиты
requests==2.31.0
